POST API auth: change URI

+++ app/src/main/java/com/techno_3_team/task_manager_google/data/LTSTViewModel.kt
@@ -70,7 +70,9 @@ class LTSTViewModel(application: Application) : AndroidViewModel(application) {
             val tasks = repository.getTasks(listId)
             repository.deleteCompletedTasks(listId)
             tasks.forEach { task ->
-                repository.deleteSubtasks(task.taskId)
+                if (task.isCompleted) {
+                    repository.deleteSubtasks(task.taskId)
+                }
             }
         }
     }
